refactor: wraps getting CMloInstanceDef logic
